REFACTOR
Now lists in settings are handled differently: prop = elem1=val1;elem2=val2

Now to get a property value you must use Settings.getProp("nameOfProp") or, if it's a list, Settings.getPropList("nameOfProp").

Defaults are stored in default.properties which is a Class Resource.

Now properties are loaded in the memory at the startup of the application instead of retrieving them when necessary (to reduce file operations).